+++ happyhouse_final_be/src/main/java/com/ssafy/happyhouse/dto/HouseInfoDto.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @ApiModel(value = "HouseInfoDto (아파트정보)", description = "아파트 정보, 거래목록, 최소거래액, 최대거래액 가진 Domain Class")
 public class HouseInfoDto {
 	@ApiModelProperty(value = "아파트코드")
-	private int aptCode;
+	private long aptCode;
 	@ApiModelProperty(value = "건축년도")
 	private int buildYear;
 	@ApiModelProperty(value = "도로명")
